+++ install/install.php
@@ -280,6 +280,7 @@ function step4 ($databasename, $newdatabasename) {
             $timezones_requirement->isValidated(),
             false,
             true,
+            false,
             false
          );
          if ($success) {
@@ -307,6 +308,7 @@ function step4 ($databasename, $newdatabasename) {
             $timezones_requirement->isValidated(),
             false,
             true,
+            false,
             false
          );
          if ($success) {
@@ -334,6 +336,7 @@ function step4 ($databasename, $newdatabasename) {
                   $timezones_requirement->isValidated(),
                   false,
                   true,
+                  false,
                   false
                );
             }